@@ -3,20 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^14.1.1",
     "axios": "^1.1.2",
+    "eslint-config-react-app": "^7.0.1",
     "formik": "^2.2.9",
     "process": "^0.11.10",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-animated-cursor": "^2.4.0",
     "react-countup": "^6.2.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-modal-video": "^1.2.8",
     "react-notifications-component": "^4.0.1",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "5.0.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "react-select": "^5.3.2",
     "react-slick": "^0.29.0",
     "sweetalert2": "^11.5.1",
